html/editing/assigning-keyboard-shortcuts/processing-model-4/accesskey.html

<!DOCTYPE html>
<meta charset=utf-8>
<title></title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<input id=i1>
<input accesskey="a b" id=i2>
<input accesskey="c dd e fff gg i" id=i3>
<input accesskey="j k ☠ l" id=i4>
<input accesskey="0 3 2 1" id=i5>
<script>
  test(function(){
    assert_equals(document.getElementById('i1').accessKey, "");
    assert_equals(document.getElementById('i1').accessKeyLabel, "");
  }, "missing accesskey attribute");

  test(function(){
    assert_equals(document.getElementById('i2').accessKey, "a b");
    assert_true(document.getElementById('i2').accessKeyLabel.contains("a b"));
  }, "accesskey attribute with single character separated by a whitespace");

  test(function(){
    assert_equals(document.getElementById('i3').accessKey, "c dd e fff gg i");
    assert_true(document.getElementById('i3').accessKeyLabel.contains("c e i"));
  }, "accesskey attribute with double and triple characters token");

  test(function(){
    assert_equals(document.getElementById('i4').accessKey, "j k ☠ l");
    assert_true(document.getElementById('i4').accessKeyLabel.contains("j k l"));
  }, "accesskey containing UTF-8");

  test(function(){
    assert_equals(document.getElementById('i5').accessKey, "0 3 2 1");
    assert_true(document.getElementById('i5').accessKeyLabel.contains("0 3 2 1"));
  }, "accesskey with digits");
</script>
